pyproject.toml
[tool.poetry]
name = "bq-flow"
version = "0.1.0"
description = "BigQuery Workflow"
authors = ["sakuv2 <loots2438@gmail.com>"]
packages = [
    { include = "bqflow", from = "src" },
]

[tool.poetry.scripts]
bqflow = 'bqflow.main:cmd'

[tool.poetry.dependencies]
python = "^3.9"
click = "^7.1.2"
pydantic = {extras = ["dotenv"], version = "^1.7.3"}
PyYAML = "^5.3.1"
abq = "^0.1.1"
sqlparse = "^0.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.4"
black = "^20.8b1"
isort = "^5.6.4"
ipykernel = "^5.4.2"
devtools = {extras = ["pygments"], version = "^0.6.1"}
mypy = "^0.800"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
exclude = '''
(
    .mypy_cache
    | .pytest_cache
    | .tox
    | .venv
    | dist
)
'''

[tool.isort]
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
